requirements.txt
# Core testing frameworks
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0

# API testing
requests==2.31.0
jsonschema==4.20.0

# UI testing
playwright==1.40.0
pytest-playwright==0.4.3

# Reporting
allure-pytest==2.13.2
pytest-json-report==1.5.0

# Data handling
faker==20.1.0
python-dotenv==1.0.0

# Excel handling
openpyxl==3.1.2
pandas==2.1.4

# Utilities
colorlog==6.8.0
python-dateutil==2.8.2